backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
gitpython==3.1.43
google-genai==0.3.0
faiss-cpu==1.8.0
numpy==1.26.4
pydantic==2.7.1
python-multipart==0.0.9
httpx==0.27.0
